Validation

+++ app/routers/user.py
@@ -4,6 +4,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from app.db.db_postgres_handler import get_session
 from app.models.models_user import UserUpdate, DeleteUserResponse, \
     FullUserResponse, UserList, Pagination, FullUserListResponse, UserSignUpRequest, UserId
+from app.routers.authintification import get_current_user
 from app.services.user import UserService
 from app.utils.utils import toFullUserResponse
 
@@ -49,16 +50,23 @@ async def create_new_user(user_data: UserSignUpRequest,
 
 @user_router.put("/{user_id}", response_model=FullUserResponse)
 async def update_existing_user(user_id: int, user_data: UserUpdate,
-                               user_service: UserService = Depends(get_user_service)) -> FullUserResponse:
+                               user_service: UserService = Depends(get_user_service),
+                               current_user: FullUserResponse = Depends(get_current_user)) -> FullUserResponse:
+    if current_user.result.user_id != user_id:
+        raise HTTPException(status_code=403, detail="Unauthorized")
     user = await user_service.update_user(user_id=user_id, user_data=user_data)
     if not user:
         raise HTTPException(status_code=404, detail="User not found")
+
     return toFullUserResponse(user)
 
 
 @user_router.delete("/{user_id}", response_model=DeleteUserResponse)
 async def delete_existing_user(user_id: int,
-                               user_service: UserService = Depends(get_user_service)) -> DeleteUserResponse:
+                               user_service: UserService = Depends(get_user_service),
+                               current_user: FullUserResponse = Depends(get_current_user)) -> DeleteUserResponse:
+    if current_user.result.user_id != user_id:
+        raise HTTPException(status_code=403, detail="Unauthorized")
     user_id = await user_service.delete_user(user_id=user_id)
     if not user_id:
         raise HTTPException(status_code=404, detail="User not found")
